Change Message-ID generation to use UUID and leak less information
A Message-ID should be globally unique. Currently mutt generates this ID
based on the current date and time, followed by ".G", followed by a letter
A to Z (A for the 1st and 27th email sent, Z for the 26th, etc.), followed
by the pid of the active mutt process, followed by "@" and the configured
fqdn.

This can lead to information being leaked as to an users email habits and
activities, which might be undesirable.

By replacing everything left of the "@" in the Message-ID with a randomly
generated UUID, we no longer include this information.