Add $crypt_sign_encrypted quadoption.
This allows automatically enabling signing for an encrypted email.
The prompt occurs after pressing send.

The requester's use-case was for $crypt_opportunistic_encrypt when
$crypt_autosign is unset.  The user doesn't want to sign _all_ their
emails, but wants to make sure oppenc encrypted emails are signed too.

It might be intellectually cleaner to add a flag for oppenc mode, or
an oppsign mode.  However the code for that would be much more
complicated.  The menus would need to be redesigned, and the
interactions with $crypt_autosign, $crypt_replysign, and
$crypt_replysignencrypted all thought through.

I don't think "oppsign" is a desired feature for the tradeoff in
complexity (I've had no requests for it), and this is a comparatively
simple addition.